backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
bcrypt
Pillow
numpy
pandas
scikit-learn
requests
Werkzeug>=3.1.0
flask-session>=0.8.0
google-auth>=2.49.0
google-auth-oauthlib>=1.3.0
google-auth-httplib2>=0.3.0
opencv-python-headless
tensorflow
tqdm
pymongo>=4.6.0
